SignGenie Backend Flask/requirements.txt
Flask==3.1.0
Flask-Cors==5.0.1
Flask-PyMongo==3.0.1
gunicorn==23.0.0
python-dotenv==1.1.0
PyJWT==2.10.1
pymongo==4.11.3
mediapipe==0.10.21
opencv-python==4.11.0.86
tensorflow==2.19.0
numpy==1.26.4
imutils==0.5.4
scikit-learn==1.6.1    
pillow==11.1.0         
requests==2.32.3       